requirements.txt
pdf2image>=1.16.3
Pillow>=8.0.0
python-dotenv>=0.19.0
pytest>=7.0.0
openai>=1.0.0
requests>=2.31.0
